lots of new functions clear erase find

+++ srcs/map.hpp
@@ -6,7 +6,7 @@
 /*   By: ahernand <ahernand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/03 19:55:43 by ahernand          #+#    #+#             */
-/*   Updated: 2022/05/14 20:50:53 by ahernand         ###   ########.fr       */
+/*   Updated: 2022/05/16 20:51:13 by ahernand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,52 +53,59 @@ namespace ft
 			size_type												_size;
 
 		public:
-		
-			/*
-			**		_______________________________ Iterators  _______________________________
-			*/
-
-			iterator begin()
-			{
-				return (ft::map_iterator<value_type, node_type>(bst_get_first(_root)));
-			}
-
-			const_iterator begin() const
-			{
-				return (ft::map_iterator<value_type, node_type>(bst_get_first(_root)));
-
-			}
-
-			iterator end()
-			{
-				return (bst_get_last(_root));
-			}
-
-			const_iterator end() const
-			{
-				return (bst_get_last(_root));
-			}
 
 
 
 
 			/*
-			**		_______________________________ Capacity  _______________________________
+			**		___________________		Iterators     ___________________
 			*/
 
-			bool				empty() const
+			iterator						begin()
+			{
+				return (ft::map_iterator<value_type, node_type>(bst_get_first(_root)));
+			}
+
+			const_iterator					begin() const
+			{
+				return (ft::map_iterator<value_type, node_type>(bst_get_first(_root)));
+
+			}
+
+			iterator						end()
+			{
+				return (bst_get_last(_root));
+			}
+
+			const_iterator					end() const
+			{
+				return (bst_get_last(_root));
+			}
+
+
+
+
+
+
+
+
+			/*
+			**		___________________		Capacity     ___________________
+			*/
+
+			bool							empty() const
 			{
 				if (_size == 0)
 					return (true);
 				return (false);
 			}
 
-			size_type			size() const
+			size_type						size() const
 			{
 				return (_size);				
 			}
 
-			size_type			max_size() const
+			size_type						max_size() const
 			{
 				return (_allocator.max_size());
 			}
@@ -106,51 +113,85 @@ namespace ft
 
 
 
+
+
+
+
 			/*
-			**		____________________________ Element Access  ____________________________
+			**		___________________		Element Access     ___________________
 			*/
 
-			mapped_type&		operator[] (const key_type& k)
+			mapped_type&					operator[] (const key_type& k)
 			{
-				node_type		*aux;
+				node_type					*aux;
 
 				aux = bst_search(_root, k);
 				if (aux == NULL)
 				{
 					T		b;
-					
+
 					insert(ft::pair<const Key, T>(k, b));
-					return (bst_search(_root, k)->data.second);
+					aux = bst_search(_root, k);
 				}
-				return(aux->data.second);
+				return (aux->data.second);
 			}
+
+
+
+
 
 
 
 
 			/*
-			**		_______________________________ Modifiers _______________________________
+			**		___________________		Modifiers     ___________________
 			*/
-			//		_________________                 Insert                 _________________
 
-			void	insert (const value_type val)
+
+
+
+			//		___________________		Insert     ___________________	
+
+
+
+
+			ft::pair<iterator, bool>		insert (const value_type &val)
 			{
+				bool					existed;
+
+				if (bst_search(_root, val.first) == NULL)
+					existed = false;
+				else
+					existed = true;
 				if (_size == 0)
 				{
 					if (_root != NULL)
 						delete _root;
 					_root = new node<value_type>(val, NULL, 0);
-					_root->left = NULL;
 					_root->right = new node<value_type>(val, _root, 1);
-					_root->parent = NULL;
 				}
 				else
 					new_node(_root, _root, val);
 				_size++;
+				return (ft::pair<iterator, bool>(iterator(bst_search(_root, val.first)), existed));
 			}
 
+
+
+
+			iterator						insert(iterator position, const value_type& val)
+			{
+				if (&position == NULL)
+					;
+				insert(val);
+				return (iterator(bst_search(_root, val.first)));
+			}
+
+
+
+
 			template <class InputIterator>
-			void insert (InputIterator first, InputIterator last)
+			void							insert (InputIterator first, InputIterator last)
 			{
 				for (; first != last; first++)
 					insert(ft::pair<const Key, T>(first->first, first->second));
@@ -159,23 +200,178 @@ namespace ft
 
 
 
-			//		_________________                 Erease                 _________________
 
-			void	erase(const key_type &val)
+
+
+
+			//		___________________		Erease     ___________________	
+
+
+
+
+			size_type						erase(const key_type &val)
 			{
-				mapped_type		mapped_aux;
+				mapped_type				mapped_aux;
+
+				if (bst_search(_root, val) == NULL)
+					return (0);				
 				_root = delete_node(_root, ft::pair<const key_type, mapped_type>(val, mapped_aux), _root);
 				--_size;
+				return (1);
 			}
 
 
 
 
+			void							erase(iterator position)
+			{
+				erase (position._ptr->data.first);
+			}
+
+
+
+
+			void							erase(iterator first, iterator last)
+			{
+				while (first != last)
+				{
+					erase (first);
+					first++;
+				}
+			}
+
+
+
+
+
+
+
+
+			//		___________________		Swap     ___________________	
+
+
+
+
+			void	swap (map& x)
+			{
+				ft::map<Key, T, Compare, Alloc>						aux;
+				
+				aux.complete_memory_copy(*this);
+				this->complete_memory_copy(x);
+				x.complete_memory_copy(aux);
+				aux._root = NULL;
+			}
+
+
+
+
+
+
+
+
+			//		___________________		Swap     ___________________	
+
+
+
+
+			void clear()
+			{
+				Key							a;
+				T							b;
+
+				_size = 0;
+				clear_tree(_root);
+				_root = new node<value_type>(ft::pair<const Key, T>(a, b), NULL, 1);
+			}
+
+
+
+
+
+
+
+
 			/*
-			**		______________________________ Operator =  ______________________________
+			**		___________________		Observers     ___________________
 			*/
 
-			map&		operator= (const map& x)
+
+
+
+			key_compare						key_comp() const
+			{
+			
+				
+			}
+
+
+
+
+
+
+
+
+			/*
+			**		___________________		Operations     ___________________
+			*/
+
+
+
+
+			iterator						find(const key_type &k)
+			{
+				node_type					*aux;
+
+				aux = bst_search(_root, k);
+				if (aux == NULL)
+					return (end());
+				return (iterator(aux));
+			}
+
+
+
+
+			const_iterator						find(const key_type &k) const
+			{
+				node_type						*aux;
+
+				aux = bst_search(_root, k);
+				if (aux == NULL)
+					return (end());
+				return (const_iterator(aux));
+			}
+
+
+
+
+
+
+
+
+			/*
+			**		___________________		Allocator     ___________________
+			*/
+			
+			
+			
+			
+			allocator_type					get_allocator() const
+			{
+				return (_allocator);				
+			}
+
+
+
+
+
+
+
+
+			/*
+			**		___________________		Operator =     ___________________
+			*/
+
+			map&							operator= (const map& x)
 			{
 				clear_tree(_root);
 				_root = NULL;
@@ -187,9 +383,16 @@ namespace ft
 
 
 
+
+
+
+
 			/*
-			**		______________________________ Cons & Dest ______________________________
+			**		___________________		Cons & Dest     ___________________
 			*/
+
+
+
 
 			explicit map (const key_compare& comp = key_compare(), const allocator_type& alloc = allocator_type())
 			{
@@ -202,23 +405,43 @@ namespace ft
 				_size = 0;
 			}
 
+
+
+
 			template <class InputIterator>
 			map (InputIterator first, InputIterator last,
 				const key_compare& comp = key_compare(),
 				const allocator_type& alloc = allocator_type())
 			{
+				_allocator = alloc;
+				_compare = comp;
 				insert(first, last);
 			}
+
+
+
 
 			map (const map& x)
 			{
 				insert(x.begin(), x.end());
 			}
 
+
+
+
 			~map()
 			{
 				clear_tree(_root);
 			}
+
+			private:
+				void	complete_memory_copy(map	&x)
+				{
+					_size = x.size();
+					_allocator = x.get_allocator();
+					_compare = x._compare; 
+					_root = x._root;
+				}
 	};
 }
